merged with sandbox

+++ ingestion_pipelines/sirivm_otp_matching_function/sirivm_otp_matching_function/sql/set_historic_matching.sql
@@ -9,7 +9,7 @@ SET time_difference = CASE
     timestamp_after_estimate = t.timestamp_after_estimate::timestamp AT TIME ZONE 'UTC',
     load_time_stamp = now()::timestamp(0)
 FROM (VALUES %s) AS t(timetable_id, time_difference, last_time_in_zone_utc, is_final_stop, journey_date, timestamp_after_estimate)
-WHERE u.timetable_id = t.timetable_id::int
+WHERE u.timetable_id = t.timetable_id::bigint
   AND date_of_journey = t.journey_date::date
   AND COALESCE(
       EXTRACT(epoch FROM (t.last_time_in_zone_utc::timestamp AT TIME ZONE 'UTC' - u.expected_departure_time)),
